+++ lib/lang/src/compile/module_compilation.rs
